Modify to work on "xwusbasp" boards, which normally do not work without special Windows drivers.
This modifies the standard USBasp firmware to fit and work on this board:

http://www.dealextreme.com/p/usb-asp-programmer-for-kk-multicopter-control-board-107402

The jumpers, incidentally, select 5V or 3.3V target power only (via the
red wire). Flashing (and USB communication) is always done at ~3.3V,
which is obtained by the drop across the two diodes.

This also makes the LED blink off as usbPoll() executes, which makes it
more useful.